requirements.txt
altgraph==0.17.2
anyio==3.6.1
astroid==2.11.6
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
bottle==0.12.21
bottle-websocket==0.2.9
certifi==2022.6.15
cffi==1.15.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.2
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
dill==0.3.5.1
Eel==0.14.0
entrypoints==0.4
executing==0.8.3
fastapi==0.78.0
fonttools==4.33.3
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
ipykernel==6.15.0
ipython==8.4.0
isort==5.10.1
jedi==0.18.1
jupyter-client==7.3.4
jupyter-core==4.10.0
kiwisolver==1.4.3
lazy-object-proxy==1.7.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.7.0
nest-asyncio==1.5.5
numpy==1.23.0
outcome==1.2.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pefile==2022.5.30
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.5.2
prompt-toolkit==3.0.30
psutil==5.9.1
psycopg2==2.9.3
pure-eval==0.2.2
pycparser==2.21
pydantic==1.9.1
Pygments==2.12.0
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.7
pylint==2.14.3
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pywin32==304
pywin32-ctypes==0.2.0
pyzmq==23.2.0
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.39
stack-data==0.3.0
starlette==0.19.1
tomli==2.0.1
tomlkit==0.11.0
tornado==6.1
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.9
uvicorn==0.18.2
wcwidth==0.2.5
whichcraft==0.6.1
wrapt==1.14.1
wsproto==1.1.0
zope.event==4.5.0
zope.interface==5.4.0
